agent/requirements.txt
pyyaml==6.0.2
mypy==1.11.2
ruff==0.6.9
bandit==1.7.9
jinja2==3.1.4
ruamel.yaml==0.18.6
